# api-gateway/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: RetailHub API
  description: |
    Event-Driven платформа диспетчеризации консультантов в ритейле.
    Клиент сканирует QR-код → создаётся заявка → консультант берёт заявку.

    ## Аутентификация
    Закрытые эндпоинты требуют JWT Bearer-токен:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Роли
    - **MANAGER** — управление магазином, отделами, консультантами, аналитика
    - **CONSULTANT** — приём и обработка заявок, управление сменами

    ## Публичные эндпоинты (без токена)
    - Сканирование QR-кода
    - Создание заявки
    - Polling статуса заявки клиентом
    - Напоминание консультанту
    - Смена консультанта
  version: 1.0.0
  contact:
    name: RetailHub Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: http://83.147.255.205:8180/api/v1
    description: Production (Direct IP)
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RequestStatus:
      type: string
      enum: [CREATED, WAITING, ESCALATED, ASSIGNED, COMPLETED, CANCELED]
      description: |
        CREATED  → WAITING   (> 3 мин без реакции, scheduler)
        WAITING  → ESCALATED (> 5 мин, SLA нарушен, scheduler)
        CREATED | WAITING | ESCALATED → ASSIGNED (консультант взял заявку)
        ASSIGNED → COMPLETED (консультант завершил)
        ASSIGNED → CREATED   (клиент нажал "Сменить консультанта")
        Любой, кроме COMPLETED/CANCELED → CANCELED (клиент отменил)

    # --- Auth ---
...
    QrCode:
      type: object
      properties:
        id:
          type: string
          format: uuid
        department_id:
          type: string
          format: uuid
        department_name:
          type: string
        token:
          type: string
          format: uuid
        scan_url:
          type: string
          format: uri
          example: https://retailhub.ru/scan/550e8400-e29b-41d4-a716-446655440000
        label:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    CreateQrCodeRequest:
      type: object
      required: [department_id]
      properties:
        department_id:
          type: string
          format: uuid
        label:
          type: string
          maxLength: 255
          example: "Стеллаж 3, ряд B"

    # --- Request ---
    ServiceRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        store_id:
          type: string
          format: uuid
        department_id:
          type: string
          format: uuid
        department_name:
          type: string
        is_escalated:
          type: boolean
          description: "True, если заявка просрочена (> 5 мин) и эскалирована менеджеру"
        assigned_user:
          $ref: "#/components/schemas/AssignedConsultant"
        status:
          $ref: "#/components/schemas/RequestStatus"
        client_session_token:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        assigned_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    AssignedConsultant:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
    CreateRequestRequest:
      type: object
      required: [qr_token]
      properties:
        qr_token:
          type: string
          format: uuid
    ClientRequestView:
      type: object
      description: Вид заявки для клиента (polling)
      properties:
        id:
          type: string
          format: uuid
        client_session_token:
          type: string
          format: uuid
          description: "Секретный токен для polling. Возвращается ТОЛЬКО при создании."
        status:
          $ref: "#/components/schemas/RequestStatus"
        department_name:
          type: string
        consultant_name:
          type: string
          description: Имя консультанта (при ASSIGNED)
        can_remind:
          type: boolean
          description: true если прошло >= 1 мин - кнопка "Напомнить"
        can_reassign:
          type: boolean
          description: true если прошло >= 3 мин - кнопка "Сменить"
        created_at:
          type: string
          format: date-time
        assigned_at:
          type: string
          format: date-time

    # --- Shift ---
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: REQUEST_ALREADY_ASSIGNED
        message:
          type: string
          example: Заявка уже взята другим консультантом
        timestamp:
          type: string
          format: date-time
    PageResponse:
      type: object
      properties:
        content:
          type: array
          items: {}
        page:
          type: integer
        size:
          type: integer
        total_elements:
          type: integer
        total_pages:
          type: integer

# PATHS
paths:
...
  /qr-codes:
    post:
      tags: [QR Codes]
      summary: Сгенерировать QR-код (MANAGER)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQrCodeRequest"
      responses:
        "201":
          description: Создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCode"

    get:
      tags: [QR Codes]
      summary: Список QR-кодов (MANAGER)
      parameters:
        - name: department_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Список
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QrCode"
  /qr-codes/{qrCodeId}:
    delete:
      tags: [QR Codes]
      summary: Деактивировать QR-код (MANAGER)
      parameters:
        - name: qrCodeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Деактивирован
    get:
      tags: [QR Codes]
      operationId: downloadQrCodeImage
      summary: Скачать PNG QR-кода для печати (MANAGER)
      description: |
        Генерирует 500×500 PNG-изображение QR-кода.
        QR-код кодирует URL: `{scanBaseUrl}/scan/{token}`.
        Браузер предложит скачать файл `qr-{id}.png`.
      security:
        - bearerAuth: []
      parameters:
        - name: qrCodeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: PNG-изображение QR-кода
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: QR-код не найден
        "400":
          description: QR-код деактивирован
  /qr-codes/scan/{token}:
    get:
      tags: [QR Codes]
      summary: Сканирование QR (публичный)
      description: Клиент попадает сюда после скана. Возвращает инфо об отделе.
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Инфо о точке
          content:
            application/json:
              schema:
                type: object
                properties:
                  department_name:
                    type: string
                  store_name:
                    type: string
                  is_valid:
                    type: boolean
        "404":
          description: QR не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # --- REQUESTS ---
  /requests:
    post:
      tags: [Requests]
      summary: Создать заявку (публичный, клиент через QR)
      description: |
        Флоу:
        1. GET /qr-codes/scan/{token} — инфо об отделе
        2. POST /requests — создание заявки
        3. GET /requests/{id}?session={token} — polling статуса
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequestRequest"
      responses:
        "201":
          description: Заявка создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientRequestView"
        "400":
          description: QR невалиден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags: [Requests]
      summary: Список заявок (MANAGER/CONSULTANT)
      description: |
        Консультант видит CREATED-заявки своих отделов.
        Менеджер видит все заявки магазина.
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/RequestStatus"
        - name: department_id
          in: query
          schema:
            type: string
            format: uuid
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Заявки
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageResponse"
                  - type: object
                    properties:
                      content:
                        type: array
                        items:
                          $ref: "#/components/schemas/ServiceRequest"
  /requests/{requestId}:
    get:
      tags: [Requests]
      summary: Статус заявки (polling клиентом, публичный)
      description: |
        Ответ включает can_remind (>= 1 мин) и can_reassign (>= 3 мин).
      security: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: session
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: client_session_token
      responses:
        "200":
          description: Статус
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientRequestView"
        "403":
          description: Неверный session-токен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /requests/{requestId}/assign:
    post:
      tags: [Requests]
      summary: Взять заявку (CONSULTANT)
      description: |
        Если уже взята — 409. Побочные эффекты:
        - Заявка → ASSIGNED, консультант → BUSY
        - Kafka → RequestAssigned
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Принята
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
        "409":
          description: Уже взята другим
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /requests/{requestId}/complete:
    post:
      tags: [Requests]
      summary: Завершить обслуживание (CONSULTANT)
      description: Заявка → COMPLETED, консультант → ACTIVE.
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Завершена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
  /requests/{requestId}/cancel:
    post:
      tags: [Requests]
      summary: Отменить заявку (клиент)
      description: Заявка → CANCELED.
      security: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: session
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Отменена
  /requests/{requestId}/remind:
    post:
      tags: [Requests]
      summary: Напомнить консультанту (клиент, >= 1 мин)
      description: Отправляет push-уведомление назначенному консультанту.
      security: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: session
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Отправлено
        "400":
          description: Рано (< 1 мин) или не ASSIGNED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /requests/{requestId}/reassign:
    post:
      tags: [Requests]
      summary: Сменить консультанта (клиент, >= 3 мин)
      description: Заявка → CREATED, предыдущий консультант → ACTIVE. Новый поиск.
      security: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: session
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: reason
          in: query
          schema:
            type: string
          description: Причина переназначения (TIMEOUT, CLIENT_REJECT)
      responses:
        "200":
          description: Возвращена в очередь
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientRequestView"
        "400":
          description: Рано (< 3 мин) или не ASSIGNED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # --- SHIFTS ---
...